vmount: another try to get abs and types right

+++ src/drivers/vmount/input_rc.cpp
@@ -120,7 +120,7 @@ bool InputRC::_read_control_data_from_subscription(ControlData &control_data, bo
 
 	// Detect a big stick movement
 	for (int i = 0; i < 3; ++i) {
-		if (fabs(_last_set_aux_values[i] - new_aux_values[i]) > 0.25f) {
+		if (fabsf(_last_set_aux_values[i] - new_aux_values[i]) > 0.25f) {
 			major_movement = true;
 		}
 	}
